+++ prefect_dbt/cli/configs/snowflake.py
@@ -8,7 +8,7 @@ except ImportError:
 
 from pydantic import Field
 
-from prefect_dbt.cli.configs.base import MissingExtrasRequireError, TargetConfigs
+from prefect_dbt.cli.configs.base import BaseTargetConfigs, MissingExtrasRequireError
 
 try:
     from prefect_snowflake.database import SnowflakeConnector
@@ -16,7 +16,7 @@ except ModuleNotFoundError as e:
     raise MissingExtrasRequireError("Snowflake") from e
 
 
-class SnowflakeTargetConfigs(TargetConfigs):
+class SnowflakeTargetConfigs(BaseTargetConfigs):
     """
     Target configs contain credentials and
     settings, specific to Snowflake.
@@ -25,9 +25,7 @@ class SnowflakeTargetConfigs(TargetConfigs):
     page.
 
     Attributes:
-        credentials: The credentials to use to authenticate; if there are
-            duplicate keys between credentials and TargetConfigs,
-            e.g. schema, an error will be raised.
+        connector: The connector to use.
 
     Examples:
         Load stored SnowflakeTargetConfigs:
@@ -56,17 +54,27 @@ class SnowflakeTargetConfigs(TargetConfigs):
             credentials=credentials,
         )
         target_configs = SnowflakeTargetConfigs(
-            connector=connector
+            connector=connector,
+            extras={"retry_on_database_errors": True},
         )
         ```
     """
 
     _block_type_name = "dbt CLI Snowflake Target Configs"
     _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/5zE9lxfzBHjw3tnEup4wWL/9a001902ed43a84c6c96d23b24622e19/dbt-bit_tm.png?h=250"  # noqa
+    _documentation_url = "https://prefecthq.github.io/prefect-dbt/cli/configs/snowflake/#prefect_dbt.cli.configs.snowflake.SnowflakeTargetConfigs"  # noqa
 
-    type: Literal["snowflake"] = "snowflake"
-    schema_: Optional[str] = Field(default=None, alias="schema")
-    connector: SnowflakeConnector
+    type: Literal["snowflake"] = Field(
+        default="snowflake", description="The type of the target configs."
+    )
+    schema_: Optional[str] = Field(
+        default=None,
+        alias="schema",
+        description="The schema to use for the target configs.",
+    )
+    connector: SnowflakeConnector = Field(
+        default=..., description="The connector to use."
+    )
 
     def get_configs(self) -> Dict[str, Any]:
         """
@@ -75,5 +83,44 @@ class SnowflakeTargetConfigs(TargetConfigs):
         Returns:
             A configs JSON.
         """
-        configs_json = super().get_configs()
+        all_configs_json = super().get_configs()
+
+        # decouple prefect-snowflake from prefect-dbt
+        # by mapping all the keys dbt snowflake accepts
+        # https://docs.getdbt.com/reference/warehouse-setups/snowflake-setup
+        rename_keys = {
+            # dbt
+            "type": "type",
+            "schema": "schema",
+            "threads": "threads",
+            # general
+            "account": "account",
+            "user": "user",
+            "role": "role",
+            "database": "database",
+            "warehouse": "warehouse",
+            # user and password
+            "password": "password",
+            # duo mfa / sso
+            "authenticator": "authenticator",
+            # key pair
+            "private_key_path": "private_key_path",
+            "private_key_passphrase": "private_key_passphrase",
+            # optional
+            "client_session_keep_alive": "client_session_keep_alive",
+            "query_tag": "query_tag",
+            "connect_retries": "connect_retries",
+            "connect_timeout": "connect_timeout",
+            "retry_on_database_errors": "retry_on_database_errors",
+            "retry_all": "retry_all",
+        }
+        configs_json = {}
+        extras = self.extras or {}
+        for key in all_configs_json.keys():
+            if key not in rename_keys and key not in extras:
+                # skip invalid keys, like fetch_size + poll_frequency_s
+                continue
+            # rename key to something dbt profile expects
+            dbt_key = rename_keys.get(key) or key
+            configs_json[dbt_key] = all_configs_json[key]
         return configs_json